requirements.txt
Pillow
devgoldyutils
re
beautifulsoup4
requests
selenium
flask
shutil